requirements.txt
streamlit==1.39.*
python-dotenv==1.0.*